Rename modid to `namereplace` (bye git history)
