+++ common/src/main/java/io/github/fifthtundrag/namereplace/NameReplace.java
@@ -1,9 +1,9 @@
-package io.github.fifthtundrag.name_replace;
+package io.github.fifthtundrag.namereplace;
 
 import com.google.gson.Gson;
-import io.github.fifthtundrag.name_replace.config.Config;
-import io.github.fifthtundrag.name_replace.config.ConfigIO;
-import io.github.fifthtundrag.name_replace.util.Platform;
+import io.github.fifthtundrag.namereplace.config.Config;
+import io.github.fifthtundrag.namereplace.config.ConfigIO;
+import io.github.fifthtundrag.namereplace.util.Platform;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -11,7 +11,7 @@ public class NameReplace {
     public static final Logger LOGGER = LoggerFactory.getLogger("NameReplace");
     public static final Gson GSON = new Gson();
 
-    public static final ConfigIO configIO = new ConfigIO(Platform.INSTANCE.getConfigDir().resolve("name_replace.json"));
+    public static final ConfigIO configIO = new ConfigIO(Platform.INSTANCE.getConfigDir().resolve("namereplace.json"));
     public static Config config;
 
     public static void init() {
