+++ common/src/main/resources/namereplace.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "io.github.fifthtundrag.name_replace.mixin",
+  "package": "io.github.fifthtundrag.namereplace.mixin",
   "compatibilityLevel": "JAVA_21",
   "mixins": [
     "ClientboundPlayerInfoUpdatePacketEntryMixin",
