+++ gradle.properties
@@ -7,8 +7,8 @@ enabled_platforms = fabric,neoforge
 fabric_loader_version=0.16.9
 # Mod Properties
 mod_version=3.0
-maven_group=io.github.fifthtundrag.name_replace
-archives_base_name=name_replace
+maven_group=io.github.fifthtundrag.namereplace
+archives_base_name=namereplace
 
 # exclude @zip for parchment version
 parchment_version=2024.11.17
